pyproject.toml
[tool.poetry]
name = "cfeconfig"
version = "0.1.0"
description = ""
authors = ["Matt Chapman <Matt@NinjitsuWeb.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
